`refactor!: restructure package organization to separate model components into distinct subpackages`
