+++ src/main/kotlin/blackjack/model/participant/Player.kt
@@ -1,4 +1,6 @@
-package blackjack.model
+package blackjack.model.participant
+
+import blackjack.model.result.PlayerResultTracker
 
 class Player(name: String) : Participant(name, PlayerResultTracker()) {
     val result: String
