+++ src/main/kotlin/blackjack/model/result/DealerResultTracker.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.result
 
 class DealerResultTracker : ResultTracker {
     private val resultMap =
