+++ src/main/kotlin/blackjack/model/result/ResultTracker.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.result
 
 interface ResultTracker {
     fun record(result: Result)
