+++ src/test/kotlin/blackjack/model/HandTest.kt
@@ -1,5 +1,7 @@
 package blackjack.model
 
+import blackjack.model.card.Card
+import blackjack.model.card.Hand
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
