+++ src/test/kotlin/blackjack/model/PlayerTest.kt
@@ -1,5 +1,7 @@
 package blackjack.model
 
+import blackjack.model.card.Card
+import blackjack.model.participant.Player
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
